changed task id to task key and removed task card component
